Move Resolution class to its own file, and add parse(...) method for initializing from various sources.
